fixed getLowByte(Integer value)
git-svn-id: http://svn.code.sf.net/p/jbidibc/code@590 87107999-b52a-42af-a34b-1e85ba8a0939

+++ trunk/jbidibc/src/main/java/org/bidib/jbidibc/utils/ByteUtils.java
@@ -114,7 +114,7 @@ public final class ByteUtils {
      */
     public static Byte getLowByte(Integer value) {
         if (value != null) {
-            byte lowByte = (byte) (value & 0xFF);
+            byte lowByte = (byte) (value.intValue() & 0xFF);
             return lowByte;
         }
         return null;
